scripts/requirements.txt
yt-dlp>=2023.12.30
python-docx>=1.1.0
